Refactor file structure; Added Socials
