+++ src/assets/icons/icon.css
@@ -1,0 +1,5 @@
+.icon-dark {
+    path {
+        fill: var(--light-text);
+    }
+}
