--- src/components/Intro.tsx
@@ -1,14 +1,0 @@
-import "../App.css";
-
-export default function About() {
-  return (
-    <>
-      <header className="intro-container">
-        <div className="intro-desc">
-          <h1>Rahul Pavithran</h1>
-          <h3>Senior Software Engineer</h3>
-        </div>
-      </header>
-    </>
-  );
-}
