+++ src/components/intro/Intro.css
@@ -1,0 +1,34 @@
+.intro-container {
+  display: flex;
+  flex-basis: 20%;
+  align-items: center;
+  justify-content: space-evenly;
+  padding: 50px 50px;
+}
+
+.intro-desc {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 10px 20px;
+  h1,
+  h3 {
+    margin: 0;
+    padding: 10px 0px;
+  }
+}
+ 
+.contact-list {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  list-style-type: none;
+
+  li {
+    width: 20%;
+    height: 20%;
+  }
+}
