+++ src/types/Icon.ts
@@ -1,0 +1,5 @@
+interface IconProps {
+    isDark: boolean,
+}
+
+export default IconProps;
